+++ packages/plugin-layout/test/routes/simple.ts
@@ -63,6 +63,7 @@ const simpleMenu = [
   { name: '欢迎', path: '/welcome', icon: 'testicon' },
   {
     name: '菜单示例',
+    path: '/test',
     children: [
       { name: '示例 1', path: '/test/1' },
       {