+++ packages/plugin-initial-state/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.8](https://github.com/umijs/plugins/compare/@umijs/plugin-initial-state@2.1.7...@umijs/plugin-initial-state@2.1.8) (2020-03-20)
+
+**Note:** Version bump only for package @umijs/plugin-initial-state
+
 ## [2.1.7](https://github.com/umijs/plugins/compare/@umijs/plugin-initial-state@2.1.6...@umijs/plugin-initial-state@2.1.7) (2020-03-18)
 
 **Note:** Version bump only for package @umijs/plugin-initial-state