+++ packages/plugin-request/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/umijs/plugins/compare/@umijs/plugin-request@2.3.1...@umijs/plugin-request@2.3.2) (2020-03-20)
+
+**Note:** Version bump only for package @umijs/plugin-request
+
 ## [2.3.1](https://github.com/umijs/plugins/compare/@umijs/plugin-request@2.3.0...@umijs/plugin-request@2.3.1) (2020-03-18)
 
 **Note:** Version bump only for package @umijs/plugin-request